feat: apply dialog to application view
