+++ src/main/kotlin/apply/ui/admin/SelectionView.kt
@@ -1,15 +1,19 @@
 package apply.ui.admin
 
 import apply.application.ApplicantService
+import apply.application.ApplicationService
+import apply.application.RecruitmentItemService
 import apply.application.RecruitmentService
 import apply.domain.applicant.Applicant
+import apply.domain.application.Application
 import com.vaadin.flow.component.Component
-import com.vaadin.flow.component.UI
+import com.vaadin.flow.component.dialog.Dialog
 import com.vaadin.flow.component.grid.Grid
 import com.vaadin.flow.component.html.H1
 import com.vaadin.flow.component.orderedlayout.FlexComponent
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout
 import com.vaadin.flow.component.orderedlayout.VerticalLayout
+import com.vaadin.flow.component.textfield.TextArea
 import com.vaadin.flow.data.renderer.ComponentRenderer
 import com.vaadin.flow.data.renderer.Renderer
 import com.vaadin.flow.router.BeforeEvent
@@ -25,13 +29,15 @@ import support.createSuccessButton
 @Route(value = "admin/selections", layout = BaseLayout::class)
 class SelectionView(
     private val applicantService: ApplicantService,
-    private val recruitmentService: RecruitmentService
+    private val applicationService: ApplicationService,
+    private val recruitmentService: RecruitmentService,
+    private val recruitmentItemService: RecruitmentItemService
 ) : VerticalLayout(), HasUrlParameter<Long> {
     private var recruitmentId: Long = 0L
 
     private fun createTitle(): Component {
         return HorizontalLayout(H1(recruitmentService.getById(recruitmentId).title)).apply {
-            setSizeFull()
+            setWidthFull()
             justifyContentMode = FlexComponent.JustifyContentMode.CENTER
         }
     }
@@ -46,7 +52,7 @@ class SelectionView(
                 // Todo: 엑셀 다운로드
             }
         ).apply {
-            setSizeFull()
+            setWidthFull()
             isSpacing = false
             justifyContentMode = FlexComponent.JustifyContentMode.BETWEEN
         }
@@ -66,12 +72,48 @@ class SelectionView(
 
     private fun createButtonRenderer(): Renderer<Applicant> {
         return ComponentRenderer<Component, Applicant> { applicant ->
-            createPrimarySmallButton("보기") {
-                UI.getCurrent().navigate(
-                    ApplicationView::class.java,
-                    "recruitments/$recruitmentId/applicants/${applicant.id}/application"
+            createPrimarySmallButton("지원서") {
+                val dialog = Dialog()
+                dialog.add(
+                    *createRecruitmentItems(
+                        applicationService.getByRecruitmentIdAndApplicantId(
+                            recruitmentId,
+                            applicant.id
+                        )
+                    )
                 )
+                dialog.width = "800px"
+                dialog.height = "90%"
+                dialog.open()
             }
+        }
+    }
+
+    private fun createRecruitmentItems(application: Application): Array<Component> {
+        val answers = application.answers
+            .items
+            .map { it.recruitmentItemId to it.contents }
+            .toMap()
+        return recruitmentItemService.findByRecruitmentId(recruitmentId)
+            .sortedBy { it.position }
+            .map {
+                createQnAs(it.title, answers[it.id] ?: throw IllegalArgumentException())
+            }.toTypedArray()
+    }
+
+    private fun createQnAs(title: String, contents: String): Component {
+        return createAnswer(title, contents).apply {
+            setWidthFull()
+            justifyContentMode = FlexComponent.JustifyContentMode.START
+        }
+    }
+
+    private fun createAnswer(title: String, answer: String): Component {
+        return TextArea(title).apply {
+            setWidthFull()
+            isReadOnly = true
+            value = answer
+            justifyContentMode = FlexComponent.JustifyContentMode.START
         }
     }
